update Game class for Board class
Separated out all of the dependency of the Game class on the Computer
class.

Responsibilities of Game
  * Rules of the game
  * Alternate turns based
  * Check gameover status
  * Return the winner (if any)

Initialy the Game class and Computer class were tightly intertwined,
depending on each other, now the Computer and the Game opperate separate
from each other. The Computer class now behaves interchangeable as a
player object in the Game.