separated computer from board dependency
Initialy the game was set up to call for a computer and a player.
I've changed the semantics so that it accounts for two players. I can
pass an instance of computer in as a player, but it is not required. All
string are severed from the computer class in order to play the game.