update jQuery to match new classes
updated the NewGame runner for all of the modification to
the classes.

Originally the game was making the alerts erupt before the winning move
was played so I reworked the runner to call it after playing the move. I
also removed the name portion of the game since nobody will ever win,
and nobody actually fills in their name anway. I had to tweek a line of
the board because my callback was not functioning properly if the board
wasn't full and a win was played.