add logic for smarter computer forks
I updated the computer to utilize the 'DoubleThreat' methods to
determine if it can cause a fork not just block it.

It does seem like a little bit of overkill since the game algorithm will
automatically create those corner forks because of a different call, and
the wall forks won't necessarily happen, but I just wanted to cover all
bases incase I were to port this logic to something like a bigger board,
or (hopefully in the near future) to a game of connect four, which might
be doable.