add two player mode
I created a two player mode to ensure that the game, player, and board
were not dependent on the computer class. I added acceptance tests to the two
player mode because I wanted to be sure that the computer isn't active or
taking turns when two player mode is selected.