optimize gameplay for mobile
I enhanced the game gameplay so that the game is playable on the phone,
landscape or vertical.

Originally the options were overlapping and you had to scroll down on the
page to see the bottom row of the game. Now when in portrait mode the
options are stacked and the game fills the bottom half of the screen. In
landscape mode the options float left and the board fills the right side
of the screen.